finished front end start backend

+++ cosmopolitan_app/backend_util/setup_cluster.sh
@@ -18,23 +18,43 @@ else
     exit 1
 fi
 
-cp "$env_file" .env
+# If the .env file exists, ask the user if they want to overwrite it.
+if [ -f .env ]; then
+    while true; do
+        read -p "The .env file already exists. Do you want to overwrite it (Y/n)? " yn
+        if [[ "$yn" =~ Y|y|^$ ]]; then
+            overwrite=true
+        fi
+        if [[ "$yn" =~ N|n ]]; then
+            overwrite=false
+        fi
+        echo "Answer Y/n."
+    done
+fi
 
-echo "Please add passwords to the .env file"
-while true; do
-    read -p "Should ${EDITOR:-vi} be used to open the file (Y/n)? " yn
-    if [[ "$yn" =~ Y|y|^$ ]]; then
-        "${EDITOR:-vi}" ".env"
-        break
-    fi
-    if [[ "$yn" =~ N|n ]]; then break; fi
-    echo "Answer Y/n."
-done
+# If the .env file does not exist, copy the appropriate .env file to the current
+# directory.
+# Ask the user to add passwords to the .env file.
+if [ "$overwrite" == true ]; then
+    cp "$env_file" .env
+
+    echo "Please add passwords to the .env file"
+    while true; do
+        read -p "Should ${EDITOR:-vi} be used to open the file (Y/n)? " yn
+        if [[ "$yn" =~ Y|y|^$ ]]; then
+            "${EDITOR:-vi}" ".env"
+            break
+        fi
+        if [[ "$yn" =~ N|n ]]; then break; fi
+        echo "Answer Y/n."
+    done
+fi
 
 source .env
 
 mkdir -p "$CLUSTER_WORK_DIR"
 
+# If the virtual environment does not exist, create it.
 if [ ! -d "$CLUSTER_PYTHON_ENV_PATH" ]; then
     python -m venv "$CLUSTER_PYTHON_ENV_PATH"
     
@@ -43,6 +63,7 @@ if [ ! -d "$CLUSTER_PYTHON_ENV_PATH" ]; then
     poetry install --no-interaction --no-ansi
 fi
 
+# Create clean up cron job
 lmod_env="BASH_ENV=/software/lmod/lmod/init/profile"
 logs="$HOME/clean_up_$1.log"
 cron_expression="00 3 * * *"
